refactor(ai): split ReindexEngineService out of ReindexEmbeddingsService (max-params 2/3)
ReindexEmbeddingsService injected 4 deps (surreal, apiKeys, embedder,
config). Extract the per-tenant re-embed engine — pagination, batched
embedMany, row updates, provider id — into ReindexEngineService
(surreal, embedder, config=batchSize). ReindexEmbeddingsService is now
the tenant orchestrator (apiKeys, engine), 2 deps. Public run() contract
and behaviour unchanged.

Part of the max-params=3 god-class split program.
